refactor(quality): polish pass — dead paths, param objects, small splits
Phase 4 (final) of the codebase-quality review — the nice-to-haves:

- GroupItemsSection: the 3-level empty-state ternary (with an unreachable
  fourth branch) becomes a named $derived with the three real cases.
- groups/[id]/members: addMember/removeMember share one loadOwnedGroup()
  opener instead of two identical load-group/404/403 blocks.
- legal/accept: all three login-redirect guards share requireUser(), so
  decline no longer loses the original redirectTo (previously hardcoded).
- hooks.server.ts: replace the stale "is this used?" TODO on handleError
  with an accurate comment (it's a standard SvelteKit hook).
- FooterComponent: the identical two-prop styling pair on all 13
  FooterLinks becomes one spread constant.
- ConversationListItem: the two nested row/title class ternaries become
  named, commented $deriveds.
- integrations: sendBatched's adjacent same-typed number params become a
  BatchPolicy object (UPDATE_POLICY/CREATE_POLICY); the duplicated
  normalizeBaseUrl + structurally identical *FetchError classes of the
  leihbackend/winbiap clients move to a shared core/http.ts (clients
  re-export, so import sites and instanceof checks are unchanged).
- TravelTimeFilter: the ORS fetch/timeout/diagnostics + geolocation
  plumbing moves to a co-located travelTimeClient.ts (the chatScroll/
  presenceHeartbeat pattern); the component keeps only UI state.
- social: the sortable/paginated trust table becomes TrustNetworkTable.svelte
  (305 → 133-line page); page-reset-on-search moves into the table.

(The sendMessage 7-positional-params finding was already fixed by the
conversations refactor #585 — it takes an options object; no change.)
