+++ src/hooks.server.ts
@@ -108,7 +108,8 @@ export const authorization: Handle = async ({ event, resolve }) => {
 
 export const handle = sequence(authentication, authorization);
 
-// TODO: Check if this is used anywhere implicitly
+// Standard SvelteKit hook: the framework calls this for every unhandled error
+// during request processing — the log line below is our only server-side trace.
 export function handleError({ error, event }): void {
 	console.error('Error occurred during request processing:', {
 		error,
